requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.7
Django==4.2.3
django-ckeditor==6.6.1
django-crispy-forms==2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.1.0
django-modeltranslation==0.18.10
Pillow==10.0.0
sqlparse==0.4.4
typing-extensions==4.7.1
